+++ test/clojure/nrepl/core_test.clj
@@ -235,8 +235,14 @@
                              :out))))
 
 (defn custom-printer
-  [value opts]
-  (format "<foo %s %s>" value (or (:sub opts) "...")))
+  ([value]
+   (custom-printer value nil))
+  ([value opts]
+   (format "<foo %s %s>" value (or (:sub opts) "..."))))
+
+(defn single-arity-printer
+  [value]
+  (format "[ %s ]" value))
 
 (def-repl-test value-printing
   (testing "bad symbol should fall back to default printer"
@@ -251,6 +257,21 @@
            (-> (message client {:op :eval
                                 :code "true"
                                 :printer `custom-printer})
+               (combine-responses)
+               (:value)))))
+  (testing "single-arity printers are supported in the absence of :print-options"
+    (is (= ["[ 42 ]"]
+           (-> (message client {:op :eval
+                                :code "42"
+                                :printer `single-arity-printer})
+               (combine-responses)
+               (:value)))))
+  (testing "empty print options are ignored"
+    (is (= ["[ 42 ]"]
+           (-> (message client {:op :eval
+                                :code "42"
+                                :printer `single-arity-printer
+                                :print-options {}})
                (combine-responses)
                (:value)))))
   (testing "options should be passed to printer"